Broke out the varmat tests into a separate file (Issue #2101)
